style(client): fix lint errors about namespaces
