+++ client/src/utils/NoStandardLibraryRequest.ts
@@ -1,0 +1,9 @@
+import { RequestType, TextDocumentIdentifier } from 'vscode-languageclient'
+
+interface NoStandardLibraryParams {
+  source: TextDocumentIdentifier
+}
+
+export const type = new RequestType<NoStandardLibraryParams, {}, void, void>(
+  'standard/noLibrary'
+)
